requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
pydantic>=2.5
python-dotenv>=1.0.0
httpx>=0.26.0
requests>=2.31
tenacity>=8.2
circuitbreaker>=1.4
Faker>=22.0
# Data serving layer
databricks-sql-connector>=3.0.0
psycopg2-binary>=2.9.9
databricks-sdk>=0.30.0
